# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Bara Sira API'
  description: "Documentation complète de l'API de Bara Sira avec Swagger"
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080'
    description: 'Serveur principal API'
paths:
...
  /api/users:
    get:
      tags:
        - Users
      operationId: users.index
      responses:
        200:
          description: Users
        403:
          $ref: '#/components/responses/Forbidden'
      security:
        -
          sanctum: []
  '/api/users/{id}':
    get:
      tags:
        - Users
      operationId: users.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
    put:
      tags:
        - Users
      operationId: users.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
    post:
      tags:
        - Users
      operationId: users.updatePost
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        200:
          description: Updated
        422:
          $ref: '#/components/responses/ValidationError'
      security:
        -
          sanctum: []
    patch:
      tags:
        - Users
      operationId: users.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
...
  '/api/service-categories/{id}':
    get:
      tags:
        - 'Service categories'
      operationId: categories.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: Category
      security: []
    put:
      tags:
        - 'Service categories'
      operationId: categories.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCategoryRequest'
      responses:
        200:
          description: Updated
      security: []
    delete:
      tags:
        - 'Service categories'
      operationId: categories.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        204:
          description: Deleted
      security: []
    patch:
      tags:
        - 'Service categories'
      operationId: categories.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCategoryRequest'
      responses:
        200:
          description: Updated
      security: []
...
  '/api/services/{id}':
    get:
      tags:
        - Services
      operationId: services.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: Service
      security: []
    put:
      tags:
        - Services
      operationId: services.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
    delete:
      tags:
        - Services
      operationId: services.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        204:
          description: Deleted
      security:
        -
          sanctum: []
    patch:
      tags:
        - Services
      operationId: services.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
...
  '/api/missions/{id}':
    get:
      tags:
        - Missions
      operationId: missions.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: Mission
      security:
        -
          sanctum: []
    put:
      tags:
        - Missions
      operationId: missions.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionUpdateRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
    delete:
      tags:
        - Missions
      operationId: missions.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        204:
          description: Deleted
      security:
        -
          sanctum: []
    patch:
      tags:
        - Missions
      operationId: missions.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionUpdateRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
...
  '/api/user-skills/{id}':
    get:
      tags:
        - 'User skills'
      operationId: skills.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: Skill
      security: []
    put:
      tags:
        - 'User skills'
      operationId: skills.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
    delete:
      tags:
        - 'User skills'
      operationId: skills.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        204:
          description: Deleted
      security: []
    patch:
      tags:
        - 'User skills'
      operationId: skills.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
...
  '/api/resumes/{id}':
    get:
      tags:
        - Resumes
      operationId: resumes.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
    put:
      tags:
        - Resumes
      operationId: resumes.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
    delete:
      tags:
        - Resumes
      operationId: resumes.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
    patch:
      tags:
        - Resumes
      operationId: resumes.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
...
  '/api/portfolio-items/{id}':
    get:
      tags:
        - Portfolio
      operationId: portfolio.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Portfolio item'
      security: []
    put:
      tags:
        - Portfolio
      operationId: portfolio.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
    delete:
      tags:
        - Portfolio
      operationId: portfolio.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        204:
          description: Deleted
      security: []
    patch:
      tags:
        - Portfolio
      operationId: portfolio.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
...
components:
  schemas:
    User:
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum:
            - client
            - prestataire
            - admin
            - superadmin
        avatar_url:
          type: string
          nullable: true
      type: object
...
    UserUpdateRequest:
      required:
        - first_name
        - last_name
        - email
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        bio:
          type: string
          maxLength: 2000
          nullable: true
        avatar:
          type: string
          format: binary
          nullable: true
        remove_avatar:
          type: boolean
      type: object
    ServiceCategoryRequest:
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 150
        description:
          type: string
          nullable: true
        icon:
          type: string
          maxLength: 255
          nullable: true
      type: object
    ServiceRequest:
      required:
        - service_category_id
        - name
        - description
        - price_min
        - price_max
      properties:
        service_category_id:
          type: integer
        name:
          type: string
          maxLength: 150
        description:
          type: string
        icon:
          type: string
          nullable: true
        price_min:
          type: number
          minimum: 0
        price_max:
          type: number
          minimum: 0
        is_active:
          type: boolean
      type: object
    MissionRequest:
      required:
        - service_id
        - title
        - description
        - city
        - address
        - date_start
      properties:
        service_id:
          type: integer
        title:
          type: string
          maxLength: 255
        description:
          type: string
        city:
          type: string
          maxLength: 50
        address:
          type: string
          maxLength: 255
        skills:
          type: array
          items:
            type: string
          maxItems: 10
        questions:
          type: array
          items:
            type: string
          maxItems: 5
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        price:
          type: number
          minimum: 0
          nullable: true
        date_start:
          type: string
          format: date-time
        date_end:
          type: string
          format: date-time
          nullable: true
      type: object
    MissionUpdateRequest:
      allOf:
        -
          $ref: '#/components/schemas/MissionRequest'
        -
          properties:
            status:
              type: string
              enum:
                - pending
                - in_progress
                - completed
                - cancelled
          type: object
...
  responses:
...
    Forbidden:
      description: 'Action forbidden'
    NotFound:
      description: 'Resource not found'
    ValidationError:
      description: 'Validation error'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
              errors:
                type: object
            type: object
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  securitySchemes:
    sanctum:
      type: apiKey
      description: 'Laravel Sanctum session cookie'
      name: laravel_session
      in: cookie
...